+++ src/App.tsx
@@ -1,70 +1,39 @@
-import { useState } from 'react'
-import './App.css'
-import { useAuthContext } from './utils/hooks';
+import { createBrowserRouter, Navigate, RouterProvider } from "react-router-dom"
+import { useAuthContext } from "./utils/hooks"
+import Home from "./pages/home/Home"
+import SignUp from "./pages/signup/SignUp"
+import Notes from "./pages/notes/Notes"
+import Dashboard from "./pages/dashboard/Dashboard"
+
 
 function App() {
-  const [name, setName] = useState('');
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const {user, signup, login, logout, loading} = useAuthContext();
-  const [error, setError] = useState<string | null>(null)
+  const {user, loading} = useAuthContext()
 
-  async function signUpUser() {
-    const resp = await signup(email, password, name);
-
-    if(resp.error) {
-      setError(resp.message)
-    } else {
-      setError(null)
-    }
-  }
-
-  async function logInUser() {
-    const resp = await login(email, password);
-
-    if(resp.error) {
-      setError(resp.message)
-    } else {
-      setError(null)
-    }
-  }
 
   if(loading) {
-    return <p>Loading...</p>
+    return <>loading...</>
   }
-  if (user) {
-    return (
-      <div>
-        <h1>Hello Everyone</h1>
-        <p>name: {user.name}</p>
-        <p>email: {user.email}</p>
-        <p>id : {user.id}</p>
-        <button onClick={logout}>Logout</button>
-        {error && <p>{error}</p>}
-      </div>
-    )
-  }
+
+  const router = createBrowserRouter([
+    {
+      path: '/',
+      element: <Home/>
+    },
+    {
+      path: 'signup',
+      element: user ? <Navigate to="/dashboard" replace /> : <SignUp />
+    },
+    {
+      path: 'dashboard',
+      element: user ? <Dashboard /> : <SignUp />
+    },
+    {
+      path: 'notes/:id',
+      element: user ? <Notes /> : <Navigate to="/signup" replace />
+    }
+  ])
   return (
-    <>
-      <div className='flex flex-col items-center gap-4 mb-8'>
-        <h1>Sign Up</h1>
-        <input value={email} placeholder='email' onChange={e => setEmail(e.target.value)} type='email'/>
-        <input value={name} placeholder='name' onChange={e => setName(e.target.value)}/>
-        <input value={password} placeholder='password' onChange={e => setPassword(e.target.value)} type='password'/>
-
-        <button className='px-4 py-1 outline outline-1 rounded' onClick={() => signUpUser()}>Sign up</button>
-      </div>
-
-      <div className='flex flex-col items-center gap-4'>
-        <h1>Login</h1>
-        <input value={email} placeholder='email' onChange={e => setEmail(e.target.value)} type='email'/>
-        <input value={password} placeholder='password' onChange={e => setPassword(e.target.value)} type='password'/>
-
-        <button className='px-4 py-1 outline outline-1 rounded' onClick={() => logInUser()}>Login</button>
-      </div>
-
-      {error && <p>{error}</p>}
-    </>
+    <RouterProvider router={router} />
   )
 }
