+++ src/pages/home/Home.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Home() {
+  return (
+    <div>Home</div>
+  )
+}
